fix(gui): harden the gate-anchor guard and split ceiling files
Review follow-ups on the #527 anchor change: pin the gate that
moved into NativeSpacesGroup (the caller needle alone pinned
only the parameter plumbing, so deleting the child's GreyOut
kept both guards green), replace the AppBarLayoutGroup anchor
needle a stray HelpButton would satisfy with the anchor's own
copy wiring, and count duplicate anchor expressions so deleting
one of a pair cannot pass a bare contains. The anchor guard
moves to its own GreyOutAnchorTests suite before the parity
file reaches the ceiling, and LayoutAppBarColorOverrides moves
out of AppBarLayoutGroup.swift (347 lines) for the same reason.
NativeSpacesGroup ignores an empty gateHelp rather than
rendering a live ? over a blank popover. ui-patterns now states
the anchor-choice rule (nearest live label scoping exactly the
gated content), classifies remote control-scoped gates (the
CrossReferenceRow model), and fixes the stale rule count.

refs #527
